+++ src/Shiny.Core/AppEvent.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+
+namespace Shiny.AppState
+{
+    public class AppEvent
+    {
+        public AppEvent(string name)
+        {
+        }
+
+
+
+    }
+}