+++ src/Shiny.Core/AppState/IAppStateDelegate.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+
+namespace Shiny
+{
+    public interface IAppStateDelegate
+    {
+        void OnStart();
+        void OnForeground();
+        void OnBackground();
+        void OnTerminate();
+    }
+}