+++ src/Shiny.Core/AppState/ServiceCollectionExtensions.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using Microsoft.Extensions.DependencyInjection;
+using Shiny.AppState;
+
+
+namespace Shiny
+{
+    public static class AppStateServiceCollectionExtensions
+    {
+        public static void AddAppState<T>(this IServiceCollection services) where T : class, IAppStateDelegate
+        {
+            if (services.IsRegistered<AppStateManager>())
+                services.AddSingleton<AppStateManager>();
+
+            services.AddSingleton<IAppStateDelegate, T>();
+        }
+    }
+}