+++ src/Shiny.Core/Platforms/Android/ShinyAndroidApplication.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Android.App;
+using Android.Content;
 using Android.Runtime;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -21,7 +22,23 @@ namespace Shiny
     }
 
 
+    //implementation "android.arch.lifecycle:extensions:1.0.0" and annotationProcessor "android.arch.lifecycle:compiler:1.0.0"
+    //class AppLifecycleListener : LifecycleObserver
+    //{
 
+    //    @OnLifecycleEvent(Lifecycle.Event.ON_START)
+    //    fun onMoveToForeground()
+    //    { // app moved to foreground
+    //    }
+
+    //    @OnLifecycleEvent(Lifecycle.Event.ON_STOP)
+    //    fun onMoveToBackground()
+    //    { // app moved to background
+    //    }
+    //}
+
+    //// register observer
+    //ProcessLifecycleOwner.get().lifecycle.addObserver(AppLifecycleListener())
     public abstract class ShinyAndroidApplication<T> : Application where T : IShinyStartup, new()
     {
         protected ShinyAndroidApplication(IntPtr javaReference, JniHandleOwnership transfer) : base(javaReference, transfer) {}
@@ -31,6 +48,23 @@ namespace Shiny
         {
             base.OnCreate();
             AndroidShinyHost.Init(this, new T(), this.OnBuildApplication);
+        }
+
+
+        public override void OnTrimMemory([GeneratedEnum] TrimMemory level)
+        {
+            if (level == TrimMemory.UiHidden || level == TrimMemory.Complete)
+            {
+                AndroidShinyHost.OnBackground();
+            }
+            base.OnTrimMemory(level);
+        }
+
+
+        public override void OnTerminate()
+        {
+            AndroidShinyHost.OnAppTerminating();
+            base.OnTerminate();
         }