+++ ChangeLog.md
@@ -1,8 +1,10 @@
 ﻿# Change Log
 
-## 1.0.1
+## 1.1.0
 
 * Shiny.Core
+    * [Feature] App State Delegate - In your shiny startup, use services.AddAppState<YourAppStateDelegate> that inherits from IAppStateDelegate.  Watch for app start, background, foreground, & termination
+    * [Enhancement][iOS][Android] Easier boilerplate setup
     * [Enhancement][Android] AndroidX support on android 10 targets - WorkManager replaces JobService under the hood
 
 * Shiny.Notifications
@@ -14,6 +16,10 @@
     * [Fix][Geofences][Android] Status check now includes GPS radio checks + permissions
     * [Fix][GPS][BREAKING] RequestAccess, WhenStatusChanged, and GetCurrentStatus all now accept GpsRequest to increase the scope/accuracy of the necessary permission checks
     * [Fix][GPS][Android] RequestAccess now checks for new Android 10 permission for ACCESS_BACKGROUND_LOCATION
+
+* Shiny.Integrations.XamarinForms
+    * [Feature] Less boilerplate to wire into XF
+    * [Feature] You can now use the DependencyService to Shiny services in your viewmodels
 
 ## 1.0.0
 Initial Release for